requirements.txt
numpy==2.0.2
opencv-python==4.12.0.88
pyzmq==26.2.0
